Test case for repeated labels.
